requirements.txt
# Core numeric
numpy>=1.24.0
scipy>=1.10.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
rembg>=2.0.50
imageio>=2.31.0

# 3D processing
trimesh>=4.0.0
open3d>=0.17.0
pymeshlab>=2022.2.post4
pygltflib>=1.16.0
pyrender>=0.1.45

# HTTP / API
httpx>=0.25.0

# Utility
tqdm>=4.65.0
rich>=13.0.0
click>=8.1.0
pydantic>=2.0.0
